Added Status API Endpoint and Implemented Make Endpoint.
Added Elastic Transcoder Pipeline Search & Create. Also added Search for Best Utterance Sequence for Dynamo (not tested).
All that is missing for a first test is the logic to send the job to ETS and pull from SNS to update the statuses and update the output url in the job.